[Silver V] Title: CD, Time: 1052 ms, Memory: 341192 KB -BaekjoonHub

+++ 백준/Silver/4158. CD/CD.java
@@ -4,27 +4,28 @@ public class Main {
 
 	public static void main(String[] args) throws Exception{
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-		StringBuilder sb = new StringBuilder();
-
+	    StringBuilder sb = new StringBuilder();
+        
 		while(true) {
 			String[] input = br.readLine().split(" ");
-			int n = Integer.parseInt(input[0]);
-			int m = Integer.parseInt(input[1]);
+			long n = Long.parseLong(input[0]);
+			long m = Long.parseLong(input[1]);
 			
 			if(n == 0 && m == 0) break;
 			
-			HashSet<Integer> set = new HashSet<>();
+			HashSet<Long> set = new HashSet<>();
 			for(int i = 0; i < n ; i ++) {
-				set.add(Integer.parseInt(br.readLine()));
+				set.add(Long.parseLong(br.readLine()));
 			}
+            
 			int cnt = 0;
 			for(int i = 0; i < m; i++) {
-				int temp = Integer.parseInt(br.readLine());
+				long temp = Long.parseLong(br.readLine());
 				if(set.contains(temp)) {
 					cnt++;
 				}
 			}
-			sb.append(cnt).append("\n");
+            sb.append(cnt).append("\n");
 		}
 		System.out.print(sb.toString());
